Compute the expected number and variance of days until the prisoner reaches freedom
A prisoner is trapped in a cell containing five doors. The first door leads to a
tunnel which returns to his cell after three days of travel. The second leads to a
tunnel which returns him to his cell after a single day of travel. The third door
leads him immediately to freedom. The fourth door leads to a tunnel that will take
him to freedom after two days of travel and the fifth door leads to a tunnel that
will take him to the beginning of the tunnel of the second door after three days of
travel. Assuming that the prisoner will always select doors 1, 2, 3, 4 and 5 with
probabilities 0.1, 0.2, 0.3, 0.1 and 0.3 respectively, simulate the system for n=20,
50, 100, 500, 1000 times to compute the expected number and variance of days
until the prisoner reaches freedom?